Split latency tests out
