+++ client/BUILD.bazel
@@ -48,3 +48,23 @@ cc_test(
         "@coroutines//:co",
     ],
 )
+
+cc_test(
+    name = "latency_test",
+    size = "small",
+    srcs = ["latency_test.cc"],
+    data = [
+        "//server:subspace_server",
+    ],
+    deps = [
+        ":subspace_client",
+        "//server",
+        "@com_google_absl//absl/flags:flag",
+        "@com_google_absl//absl/flags:parse",
+        "@com_google_absl//absl/hash:hash_testing",
+        "@com_google_absl//absl/status",
+        "@com_google_absl//absl/status:statusor",
+        "@com_google_googletest//:gtest",
+        "@coroutines//:co",
+    ],
+)
